video_call_app/requirements.txt
flask==2.3.3
flask-socketio==5.3.4
python-socketio==5.8.0
python-engineio==4.5.1
eventlet==0.33.3
pyOpenSSL==23.2.0
